refactor(middleware): split middleware.rs into a folder + relocate token_bucket
Combines Finding 2 (relocate `worker/token_bucket.rs`) and a long-overdue
break-up of `middleware.rs` from `.claude/plans/2026-04-14-worker-module-followup-cleanup.md`
into a single PR. Pure structural refactor — zero behavior change, every
external import path stays valid via re-exports in `middleware/mod.rs`.

Why one PR
- `middleware.rs` was importing `crate::worker::token_bucket::TokenBucket`
  via `pub use`, so the cleanest moment to relocate token_bucket is when
  middleware itself becomes a folder. Doing the two changes in separate
  PRs would create a transient `middleware.rs` → `middleware/mod.rs` →
  `middleware.rs` flip that nobody needs.

Layout
Before:
  model_gateway/src/middleware.rs           (1110 lines, 8 concerns)
  model_gateway/src/worker/token_bucket.rs

After:
  model_gateway/src/middleware/mod.rs              (re-exports for back-compat)
  model_gateway/src/middleware/auth.rs             (AuthConfig + auth_middleware)
  model_gateway/src/middleware/concurrency.rs      (TokenGuardBody, QueuedRequest,
                                                    QueueProcessor, ConcurrencyLimiter,
                                                    concurrency_limit_middleware)
  model_gateway/src/middleware/logging.rs          (RequestSpan, RequestLogger,
                                                    ResponseLogger, create_logging_layer)
  model_gateway/src/middleware/metrics.rs          (HttpMetricsLayer,
                                                    HttpMetricsMiddleware,
                                                    normalize_path_for_metrics +
                                                    is_dynamic_id + tests)
  model_gateway/src/middleware/request_id.rs       (RequestIdLayer,
                                                    RequestIdMiddleware,
                                                    generate_request_id, RequestId
                                                    re-export from smg_auth)
  model_gateway/src/middleware/storage_context.rs  (storage_context_middleware +
                                                    helpers + tests)
  model_gateway/src/middleware/token_bucket.rs     (relocated from worker/, git
                                                    rename so history follows)
  model_gateway/src/middleware/wasm.rs             (wasm_middleware)

What changed
- `model_gateway/src/middleware.rs`: deleted (1110 lines distributed across
  the 8 files above, byte-for-byte identical except for per-file imports).
- `model_gateway/src/middleware/mod.rs`: new — registers all submodules and
  re-exports the public surface (`AuthConfig`, `auth_middleware`,
  `ConcurrencyLimiter`, `QueueProcessor`, `QueuedRequest`, `TokenGuardBody`,
  `concurrency_limit_middleware`, `create_logging_layer`, `RequestLogger`,
  `RequestSpan`, `ResponseLogger`, `HttpMetricsLayer`, `HttpMetricsMiddleware`,
  `RequestId`, `RequestIdLayer`, `RequestIdMiddleware`,
  `storage_context_middleware`, `TokenBucket`, `wasm_middleware`).
- `model_gateway/src/worker/token_bucket.rs` → `model_gateway/src/middleware/token_bucket.rs`:
  git rename, file content unchanged. Tracks as a rename in the diff.
- `model_gateway/src/worker/mod.rs`: drops `pub mod token_bucket;`.
- Internal cross-module wiring inside the new folder:
  * `concurrency.rs` imports `super::token_bucket::TokenBucket`.
  * `logging.rs` imports `super::metrics::normalize_path_for_metrics` and
    `super::request_id::RequestId`.
  * `wasm.rs` imports `super::request_id::{generate_request_id, RequestId}`.
  * `generate_request_id` is `pub(super)` so `wasm.rs` can use it without
    leaking it past the module boundary.
  * `normalize_path_for_metrics` is `pub(super)` so `logging.rs` can use it.
- Test modules followed their code: `build_storage_request_context_*`
  tests live in `storage_context.rs`; the `normalize_path_*` and
  `is_dynamic_id` tests live in `metrics.rs`.

Public surface preserved
- `crate::middleware::TokenBucket` still resolves (now via
  `middleware/mod.rs:pub use token_bucket::TokenBucket`), so the existing
  callers in `app_context.rs:19` and `service_discovery.rs:1240` are
  untouched.
- Every `crate::middleware::{AuthConfig, QueuedRequest, …}` import in
  `server.rs` still resolves via the re-exports.
- `RequestId` re-export from `smg_auth` is preserved (now in
  `middleware/request_id.rs`).
- Net: no caller outside `worker/mod.rs` needs an import change.

What this unblocks
- `worker/` no longer hosts a primitive that has zero coupling to the
  Worker abstraction — Finding 2 closed.
- `middleware.rs` is no longer a 1110-line bag of concerns; future
  middleware additions can live in their own files instead of growing
  the monolith.
- The folder layout sets a precedent for the next reviewer: one concern
  per file, each file imports only what it uses.

What this does NOT do
- No new middleware trait. Tower's `Layer`/`Service` is already the
  ecosystem-standard abstraction; introducing a project-local trait
  would just be a wrapper.
- No behavior change. No test changes (tests moved alongside their
  code, but their assertions are identical).
- No removal of public API. Every name that was public before is still
  public.

Verification
- cargo check -p smg — clean.
- cargo clippy -p smg --all-targets -- -D warnings — clean.
- cargo clippy -p smg-golang --all-targets -- -D warnings — clean.
- cargo check -p smg-python — clean.
- cargo test -p smg --lib — 546 passed; 0 failed; 4 ignored.
- make fmt — clean.
- rustup run nightly cargo fmt -- --check — EXIT=0 (matches CI).

Refs: Finding 2 in .claude/plans/2026-04-14-worker-module-followup-cleanup.md
