+++ model_gateway/src/middleware/request_id.rs
@@ -1,0 +1,129 @@
+//! Request ID middleware — pulls an existing request ID from configured
+//! headers or generates an OpenAI-compatible one, stores it in extensions,
+//! and echoes it back via `x-request-id`.
+
+use std::{
+    pin::Pin,
+    sync::Arc,
+    task::{Context, Poll},
+};
+
+use axum::{extract::Request, http::HeaderValue, response::Response};
+use rand::Rng;
+// Re-export RequestId from auth crate for backward compatibility
+pub use smg_auth::RequestId;
+use tower::{Layer, Service};
+
+/// Alphanumeric characters for request ID generation (as bytes for O(1) indexing)
+const REQUEST_ID_CHARS: &[u8] = b"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
+
+/// Generate OpenAI-compatible request ID based on endpoint.
+pub(super) fn generate_request_id(path: &str) -> String {
+    let prefix = if path.contains("/chat/completions") {
+        "chatcmpl-"
+    } else if path.contains("/completions") {
+        "cmpl-"
+    } else if path.contains("/generate") {
+        "gnt-"
+    } else if path.contains("/responses") {
+        "resp-"
+    } else {
+        "req-"
+    };
+
+    // Generate a random string similar to OpenAI's format
+    // Use byte array indexing (O(1)) instead of chars().nth() (O(n))
+    let mut rng = rand::rng();
+    let random_part: String = (0..24)
+        .map(|_| {
+            let idx = rng.random_range(0..REQUEST_ID_CHARS.len());
+            REQUEST_ID_CHARS[idx] as char
+        })
+        .collect();
+
+    format!("{prefix}{random_part}")
+}
+
+/// Tower Layer for request ID middleware
+#[derive(Clone)]
+pub struct RequestIdLayer {
+    headers: Arc<Vec<String>>,
+}
+
+impl RequestIdLayer {
+    pub fn new(headers: Vec<String>) -> Self {
+        Self {
+            headers: Arc::new(headers),
+        }
+    }
+}
+
+impl<S> Layer<S> for RequestIdLayer {
+    type Service = RequestIdMiddleware<S>;
+
+    fn layer(&self, inner: S) -> Self::Service {
+        RequestIdMiddleware {
+            inner,
+            headers: self.headers.clone(),
+        }
+    }
+}
+
+/// Tower Service for request ID middleware
+#[derive(Clone)]
+pub struct RequestIdMiddleware<S> {
+    inner: S,
+    headers: Arc<Vec<String>>,
+}
+
+impl<S> Service<Request> for RequestIdMiddleware<S>
+where
+    S: Service<Request, Response = Response> + Send + 'static,
+    S::Future: Send + 'static,
+{
+    type Response = S::Response;
+    type Error = S::Error;
+    type Future =
+        Pin<Box<dyn std::future::Future<Output = Result<Self::Response, Self::Error>> + Send>>;
+
+    fn poll_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        self.inner.poll_ready(cx)
+    }
+
+    fn call(&mut self, mut req: Request) -> Self::Future {
+        let headers = self.headers.clone();
+
+        // Extract request ID from headers or generate new one
+        let mut request_id = None;
+
+        for header_name in headers.iter() {
+            if let Some(header_value) = req.headers().get(header_name) {
+                if let Ok(value) = header_value.to_str() {
+                    request_id = Some(value.to_string());
+                    break;
+                }
+            }
+        }
+
+        let request_id = request_id.unwrap_or_else(|| generate_request_id(req.uri().path()));
+
+        // Insert request ID into request extensions for other middleware/handlers to use
+        req.extensions_mut().insert(RequestId(request_id.clone()));
+
+        // Call the inner service
+        let future = self.inner.call(req);
+
+        Box::pin(async move {
+            let mut response = future.await?;
+
+            // Add request ID to response headers
+            response.headers_mut().insert(
+                "x-request-id",
+                HeaderValue::from_str(&request_id)
+                    .unwrap_or_else(|_| HeaderValue::from_static("invalid-request-id")),
+            );
+
+            Ok(response)
+        })
+    }
+}
