+++ model_gateway/src/worker/mod.rs
@@ -13,7 +13,6 @@ pub mod monitor;
 pub mod registry;
 pub mod resilience;
 pub mod service;
-pub mod token_bucket;
 // FIXME: worker.rs is a 1800-line monolith containing the Worker trait,
 // BasicWorker impl, HealthChecker, WorkerType, ConnectionMode, and more.
 // Break it apart into focused modules (e.g. health_checker.rs, types.rs).
